% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCocluster.R
\name{binarydata}
\alias{binarydata}
\title{Simulated Binary Data-set}
\format{A data matrix with 1000 rows and 100 columns.}
\description{
It is a binary data-set simulated using Bernoulli distribution. It consist of two clusters in rows and three 
clusters in columns.
}
\examples{
  data(binarydata)
}
\keyword{datasets}
