\name{BinaryOptions-class}
\alias{BinaryOptions-class}
\title{Binary input/output options}
\description{
  This class contains all the input options as well as the
  estimated paramters for Binary data-set. It inherits from
  base class \code{\linkS4class{CommonOptions}}
}
\details{
  \describe{ \item{classmean:}{The mean value of each
  co-cluster.} \item{classdispersion:}{The dispersion of
  each co-cluster.} }
}

