% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionclasses.R
\docType{class}
\name{CategoricalOptions-class}
\alias{CategoricalOptions-class}
\title{Categorical input/output options}
\description{
This class contains all the input options as well as the estimated paramters for categorical data-set. It inherits
from base class \code{\linkS4class{CommonOptions}}. The class contains following output parameters given in 'Details' along
with the parameters in base class.
}
\details{
\describe{
\item{classmean:}{The categorical distribution of each co-cluster}
\item{ICLvalue: }{Integrated complete likelihood}
}
}
