\name{ContinuousOptions-class}
\alias{ContinuousOptions-class}
\title{Continuous input/output options}
\description{
  This class contains all the input options as well as the
  estimated paramters for Continuous data-sets. It inherits
  from base class \code{\linkS4class{CommonOptions}}. The
  class contains following output parameters given in
  'Details' along with the parameters in base class.
}
\details{
  \describe{ \item{classmean:}{The mean value of each
  co-cluster.} \item{classvariance:}{The variance of each
  co-cluster.} }
}

