\name{[}
\alias{[}
\alias{[,BinaryOptions-method}
\alias{[,CategoricalOptions-method}
\alias{[,ContingencyOptions-method}
\alias{[,ContinuousOptions-method}
\alias{[,strategy-method}
\title{Getter method for Rcoclust output}
\arguments{
  \item{x}{object from which to extract element(s) or in
  which to replace element(s).}

  \item{i}{the name of the element we want to extract or
  replace.}

  \item{j}{if the element designing by i is complex, j
  specifying elements to extract or replace.}
}
\description{
  This is overloading of square braces to extract values of
  various slots of the output from \code{\link{cocluster}}.
}

