\name{strategy-class}
\alias{strategy-class}
\title{strategy class}
\description{
  This class contains all the input parameters to run
  coclustering.
}
\details{
  \describe{ \item{algo: }{Algorithm to be use for
  co-clustering.} \item{stopcriteria: }{Stopping criteria
  used to stop the algorithm.} \item{initmethod: }{Method
  to initialize model paramteres.} \item{nbinititerations:
  }{Number of global iterations while running
  initialization.} \item{initepsilon: }{Tolerance value
  used while initialization.} \item{nbiterations_int:
  }{Number of iterations for internal E-step.}
  \item{epsilon_int: }{Tolerance value for internal
  E-step.} \item{nbtry: }{Number of tries.} \item{nbxem:
  }{Number of xem iterations.} \item{nbiterationsxem:
  }{Number of EM iterations used during xem.}
  \item{nbiterationsXEM: }{Number of EM iterations used
  during XEM.} \item{epsilonxem: }{Tolerance value used
  during xem.} \item{epsilonXEM: }{Tolerance value used
  during XEM.} }
}

