% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionclasses.R
\docType{class}
\name{BinaryOptions-class}
\alias{BinaryOptions-class}
\title{Binary input/output options}
\description{
This class contains all the input options as well as the estimated paramters for Binary data-set. It inherits
from base class \code{\linkS4class{CommonOptions}}. The class contains following output parameters given in 'Details' along
with the parameters in base class.
}
\details{
\describe{
\item{classmean:}{The mean value of each co-cluster.}
\item{classdispersion:}{The dispersion of each co-cluster.}
\item{ICLvalue: }{Integrated complete likelihood}
}
}
