% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclusterStrategy.R, R/optionclasses.R
\docType{methods}
\name{[,strategy-method}
\alias{[,strategy-method}
\alias{[}
\alias{[,BinaryOptions-method}
\alias{[,ContingencyOptions-method}
\alias{[,ContinuousOptions-method}
\alias{[,CategoricalOptions-method}
\title{Getter method for blockcluster output}
\usage{
\S4method{[}{strategy}(x, i, j, drop)

\S4method{[}{BinaryOptions}(x, i, j, drop)

\S4method{[}{ContingencyOptions}(x, i, j, drop)

\S4method{[}{ContinuousOptions}(x, i, j, drop)

\S4method{[}{CategoricalOptions}(x, i, j, drop)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i}{the name of the element we want to extract or replace.}

\item{j}{if the element designing by i is complex, j specifying elements to extract or replace.}

\item{drop}{not used}
}
\description{
This is overloading of square braces to extract values of various slots of
the output from \code{\link{coclusterBinary}},
\code{\link{coclusterCategorical}}, \code{\link{coclusterContingency}},
\code{\link{coclusterContinuous}}.
}
