% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCocluster.R, R/optionclasses.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,BinaryOptions-method}
\alias{plot,ContingencyOptions-method}
\alias{plot,ContinuousOptions-method}
\alias{plot,CategoricalOptions-method}
\title{Plot function.}
\usage{
\S4method{plot}{BinaryOptions}(x, y, ...)

\S4method{plot}{ContingencyOptions}(x, y, ...)

\S4method{plot}{ContinuousOptions}(x, y, ...)

\S4method{plot}{CategoricalOptions}(x, y, ...)
}
\arguments{
\item{x}{output object from \code{\link{coclusterBinary}},
\code{\link{coclusterCategorical}}, \code{\link{coclusterContingency}},
\code{\link{coclusterContinuous}}.}

\item{y}{Ignored}

\item{...}{Additional argument(s). Currently we support two additional argument.
"asp": If this is set to TRUE the original aspect ratio is conserved. By
default "asp" is FALSE. "type" : This is the type of plot which is either
"cocluster" or "distribution". The corresponding plots are Co-clustered data
and distributions and mixture densities for Co-clusters respectively.
Default is "cocluster" plot.}
}
\description{
This function plot the original and Co-clustered data-sets.
}
