% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclusterStrategy.R, R/RCocluster.R,
%   R/optionclasses.R
\docType{methods}
\name{summary,strategy-method}
\alias{summary,strategy-method}
\alias{summary}
\alias{summary,BinaryOptions-method}
\alias{summary,ContingencyOptions-method}
\alias{summary,ContinuousOptions-method}
\alias{summary,CategoricalOptions-method}
\title{Summary function.}
\usage{
\S4method{summary}{strategy}(object, ...)

\S4method{summary}{BinaryOptions}(object, ...)

\S4method{summary}{ContingencyOptions}(object, ...)

\S4method{summary}{ContinuousOptions}(object, ...)

\S4method{summary}{CategoricalOptions}(object, ...)
}
\arguments{
\item{object}{output object from \code{\link{coclusterBinary}},
\code{\link{coclusterCategorical}}, \code{\link{coclusterContingency}},
\code{\link{coclusterContinuous}}.}

\item{...}{Additional argument(s) . Currently there is no additional arguments.}
}
\description{
This function gives the summary of output from \code{coclusterBinary},
\code{coclusterCategorical}, \code{coclusterContingency},
\code{coclusterContinuous}.
}
