\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,BinaryOptions-method}
\alias{summary,ContingencyOptions-method}
\alias{summary,ContinuousOptions-method}
\alias{summary,strategy-method}
\title{Summary function.}
\arguments{
  \item{object}{output object from
  \code{\link{cocluster}}.}
}
\description{
  This funtion gives the summary of output from
  \code{cocluster}.
}

